Move rule handlers into sub-namespace
