+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/Tree/Dependencies/Subscriptions/RuleHandlers/ComRuleHandler.cs
@@ -3,35 +3,34 @@
 using System.Collections.Immutable;
 using System.ComponentModel.Composition;
 
-using Microsoft.VisualStudio.Imaging;
 using Microsoft.VisualStudio.Imaging.Interop;
 using Microsoft.VisualStudio.ProjectSystem.VS.Tree.Dependencies.CrossTarget;
 using Microsoft.VisualStudio.ProjectSystem.VS.Tree.Dependencies.Models;
 
-namespace Microsoft.VisualStudio.ProjectSystem.VS.Tree.Dependencies.Subscriptions
+namespace Microsoft.VisualStudio.ProjectSystem.VS.Tree.Dependencies.Subscriptions.RuleHandlers
 {
     [Export(DependencyRulesSubscriber.DependencyRulesSubscriberContract,
             typeof(IDependenciesRuleHandler))]
     [Export(typeof(IProjectDependenciesSubTreeProvider))]
     [AppliesTo(ProjectCapability.DependenciesTree)]
-    internal class AnalyzerRuleHandler : DependenciesRuleHandlerBase
+    internal class ComRuleHandler : DependenciesRuleHandlerBase
     {
-        public const string ProviderTypeString = "AnalyzerDependency";
+        public const string ProviderTypeString = "ComDependency";
 
         private static readonly DependencyIconSet s_iconSet = new DependencyIconSet(
-            icon: KnownMonikers.CodeInformation,
-            expandedIcon: KnownMonikers.CodeInformation,
-            unresolvedIcon: ManagedImageMonikers.CodeInformationWarning,
-            unresolvedExpandedIcon: ManagedImageMonikers.CodeInformationWarning);
+            icon: ManagedImageMonikers.Component,
+            expandedIcon: ManagedImageMonikers.Component,
+            unresolvedIcon: ManagedImageMonikers.ComponentWarning,
+            unresolvedExpandedIcon: ManagedImageMonikers.ComponentWarning);
 
         private static readonly SubTreeRootDependencyModel s_rootModel = new SubTreeRootDependencyModel(
             ProviderTypeString,
-            VSResources.AnalyzersNodeName,
+            VSResources.ComNodeName,
             s_iconSet,
-            DependencyTreeFlags.AnalyzerSubTreeRootNodeFlags);
+            DependencyTreeFlags.ComSubTreeRootNodeFlags);
 
-        public AnalyzerRuleHandler()
-            : base(AnalyzerReference.SchemaName, ResolvedAnalyzerReference.SchemaName)
+        public ComRuleHandler()
+            : base(ComReference.SchemaName, ResolvedCOMReference.SchemaName)
         {
         }
 
@@ -46,7 +45,7 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.Tree.Dependencies.Subscription
             bool isImplicit,
             IImmutableDictionary<string, string> properties)
         {
-            return new AnalyzerDependencyModel(
+            return new ComDependencyModel(
                 path,
                 originalItemSpec,
                 resolved,
@@ -56,7 +55,7 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.Tree.Dependencies.Subscription
 
         public override ImageMoniker GetImplicitIcon()
         {
-            return ManagedImageMonikers.CodeInformationPrivate;
+            return ManagedImageMonikers.ComponentPrivate;
         }
     }
 }
